+++ app/src/main/java/com/battlelancer/seriesguide/shows/ShowsDistillationFragment.kt
@@ -10,21 +10,20 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.appcompat.app.AppCompatDialogFragment
 import androidx.compose.ui.platform.ViewCompositionStrategy
-import androidx.core.content.edit
 import androidx.core.view.isGone
 import androidx.fragment.app.DialogFragment
 import androidx.fragment.app.FragmentManager
+import androidx.fragment.app.activityViewModels
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
-import androidx.preference.PreferenceManager
 import com.battlelancer.seriesguide.R
-import com.battlelancer.seriesguide.appwidget.ListWidgetProvider
 import com.battlelancer.seriesguide.databinding.DialogShowsDistillationBinding
 import com.battlelancer.seriesguide.settings.AdvancedSettings
 import com.battlelancer.seriesguide.settings.DisplaySettings
-import com.battlelancer.seriesguide.shows.ShowsDistillationSettings.ShowFilter
+import com.battlelancer.seriesguide.shows.ShowsDistillationSettings.ShowFilters
+import com.battlelancer.seriesguide.shows.ShowsDistillationSettings.ShowSortOrder
 import com.battlelancer.seriesguide.streaming.SgWatchProvider
 import com.battlelancer.seriesguide.streaming.StreamingSearchInfoDialog
 import com.battlelancer.seriesguide.ui.dialogs.SingleChoiceDialogFragment
@@ -36,6 +35,7 @@ import kotlinx.coroutines.launch
 
 class ShowsDistillationFragment : AppCompatDialogFragment() {
 
+    private val activityModel: ShowsActivityViewModel by activityViewModels()
     private val model: ShowsDistillationViewModel by viewModels()
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -53,8 +53,8 @@ class ShowsDistillationFragment : AppCompatDialogFragment() {
 
         val binding = DialogShowsDistillationBinding.inflate(inflater, container, false)
 
-        val initialShowFilter = ShowFilter.fromSettings(requireContext())
-        val initialShowSortOrder = SortShowsView.ShowSortOrder.fromSettings(requireContext())
+        val initialShowFilters = ShowFilters.fromSettings(requireContext())
+        val initialShowSortOrder = ShowSortOrder.fromSettings(requireContext())
 
         binding.apply {
             filterShowsView.isGone = false
@@ -79,7 +79,7 @@ class ShowsDistillationFragment : AppCompatDialogFragment() {
 
             filterShowsView.apply {
                 setInitialFilter(
-                    initialShowFilter,
+                    initialShowFilters,
                     DisplaySettings.isNoReleasedEpisodes(context)
                 )
                 setFilterListener(filterListener)
@@ -110,11 +110,8 @@ class ShowsDistillationFragment : AppCompatDialogFragment() {
             // Display tab icon with indicator if general filter is active
             lifecycleScope.launch {
                 repeatOnLifecycle(Lifecycle.State.STARTED) {
-                    ShowsDistillationSettings.showFilter.collect { filters ->
-                        filters?.let {
-                            val isFiltering = filters.isAnyFilterEnabled()
-                            tabLayoutShowsDistillation.setFilterIsActiveIcon(0, isFiltering)
-                        }
+                    activityModel.showsDistillationSettings.showFilters.collect {
+                        tabLayoutShowsDistillation.setFilterIsActiveIcon(0, it.isAnyFilterEnabled())
                     }
                 }
             }
@@ -146,8 +143,8 @@ class ShowsDistillationFragment : AppCompatDialogFragment() {
     }
 
     private val filterListener = object : FilterShowsView.FilterListener {
-        override fun onFilterUpdate(filter: ShowFilter) {
-            ShowsDistillationSettings.saveFilter(requireContext(), filter)
+        override fun onFilterUpdate(filters: ShowFilters) {
+            activityModel.showsDistillationSettings.saveFilters(filters)
         }
 
         override fun onConfigureUpcomingRangeClick() {
@@ -190,29 +187,9 @@ class ShowsDistillationFragment : AppCompatDialogFragment() {
     }
 
     private val sortOrderListener = object : SortShowsView.SortOrderListener {
-        override fun onSortOrderUpdate(showSortOrder: SortShowsView.ShowSortOrder) {
-            // save new sort order to preferences
-            PreferenceManager.getDefaultSharedPreferences(requireContext()).edit {
-                putInt(ShowsDistillationSettings.KEY_SORT_ORDER, showSortOrder.sortOrderId)
-                putBoolean(
-                    ShowsDistillationSettings.KEY_SORT_FAVORITES_FIRST,
-                    showSortOrder.isSortFavoritesFirst
-                )
-                putBoolean(
-                    DisplaySettings.KEY_SORT_IGNORE_ARTICLE,
-                    showSortOrder.isSortIgnoreArticles
-                )
-            }
-
-            // broadcast new sort order
-            ShowsDistillationSettings.sortOrder.value = showSortOrder
-
-            if (showSortOrder.changedIgnoreArticles) {
-                // refresh all list widgets
-                ListWidgetProvider.notifyDataChanged(context!!)
-            }
+        override fun onSortOrderUpdate(showSortOrder: ShowSortOrder) {
+            activityModel.showsDistillationSettings.saveSortOrder(showSortOrder)
         }
-
     }
 
     companion object {